REFACTOR: ajustando pasta routes e atualizando documentação
